Database Vectorizer/requirements.txt
pymongo
python-dotenv
sentence_transformers